src/orders/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-property==0.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
confluent-kafka==2.12.1
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.120.2
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jwcrypto==1.5.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pymongo==4.15.3
python-dotenv==1.2.1
python-keycloak==5.8.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.2
sentry-sdk==2.43.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.49.1
structlog==25.5.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
pytest==8.4.2
